+++ dpws/src/test/java/org/somda/sdc/dpws/soap/wsdiscovery/WsDiscoveryUtilTest.java
@@ -6,6 +6,7 @@ import org.somda.sdc.dpws.DpwsTest;
 
 import java.util.List;
 
+import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 class WsDiscoveryUtilTest extends DpwsTest {
@@ -25,5 +26,39 @@ class WsDiscoveryUtilTest extends DpwsTest {
         var subset = List.of("http://b.de");
 
         assertTrue(wsDiscoveryUtil.isScopesMatching(superset, subset, MatchBy.STRCMP0));
+    }
+
+    @Test
+    void isScopesMatchingRfc3986Test() {
+        var superset = List.of("http://a.de/abc/d");
+
+        // equals
+        assertTrue(doesMatch(superset, "http://a.de/abc/d"));
+        // case insensitive schema
+        assertTrue(doesMatch(superset, "HTTP://a.de/abc/d"));
+        // case insensitive authority
+        assertTrue(doesMatch(superset, "http://A.dE/abc/d"));
+        // match by segment
+        assertTrue(doesMatch(superset, "http://a.de/abc"));
+        // match by segment, trailing slash ignored
+        assertTrue(doesMatch(superset, "http://a.de/abc/"));
+        // match if subset has no segments
+        assertTrue(doesMatch(superset, "http://a.de"));
+
+        // case sensitive segment
+        assertFalse(doesMatch(superset, "http://a.de/Abc/d"));
+        // encoded URI doesn't match even if it would after decoding (http://a.de/abc/d)
+        assertFalse(doesMatch(superset, "http%3A%2F%2Fa.de%2Fabc%2Fd"));
+        // doesn't match if only part of segment matches
+        assertFalse(doesMatch(superset, "http://a.de/ab"));
+        // doesn't match if subset has more segment than superset
+        assertFalse(doesMatch(superset, "http://a.de/abc/d/d"));
+        // doesn't match if subset has '.' or '..' in path
+        assertFalse(doesMatch(superset, "http://a.de/abc/./d"));
+        assertFalse(doesMatch(superset, "http://a.de/abc/../d"));
+    }
+
+    private boolean doesMatch(List<String> superset, String s) {
+        return wsDiscoveryUtil.isScopesMatching(superset, List.of(s), MatchBy.RFC3986);
     }
 }